pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ai-research>=0.1.36",
    "ipykernel>=7.2.0",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pip>=26.0.1",
    "python-dotenv>=1.2.1",
    "scipy>=1.15.3",
    "skrl>=1.4.3",
]
